Merge from develop to release branch

+++ partner-onboarder/install.sh
@@ -17,11 +17,11 @@ if [ -z "$flag" ]; then
 fi
 ENABLE_INSECURE=''
 if [ "$flag" = "n" ]; then
-  ENABLE_INSECURE='--set onboarding.enableInsecure=true';
+  ENABLE_INSECURE='--set onboarding.configmaps.onboarding.ENABLE_INSECURE=true';
 fi
 
 NS=esignet
-CHART_VERSION=12.0.1-B3
+CHART_VERSION=12.0.2
 
 echo Create $NS namespace
 kubectl create ns $NS
@@ -38,38 +38,56 @@ function installing_onboarder() {
     kubectl -n $NS --ignore-not-found=true delete cm s3
     sed -i 's/\r$//' copy_cm.sh
     ./copy_cm.sh
+    kubectl -n $NS delete cm --ignore-not-found=true onboarding
 
     echo Copy secrets
     sed -i 's/\r$//' copy_secrets.sh
     ./copy_secrets.sh
 
+    read -p "Provide onboarder bucket name : " s3_bucket
+    if [[ -z $s3_bucket ]]; then
+      echo "s3_bucket not provided; EXITING;";
+      exit 1;
+    fi
+    if [[ $s3_bucket == *[' !@#$%^&*()+']* ]]; then
+      echo "s3_bucket should not contain spaces / any special character; EXITING";
+      exit 1;
+    fi
+    read -p "Provide onboarder s3 bucket region : " s3_region
+    if [[ $s3_region == *[' !@#$%^&*()+']* ]]; then
+      echo "s3_region should not contain spaces / any special character; EXITING";
+      exit 1;
+    fi
+
+    read -p "Provide S3 URL : " s3_url
+    if [[ -z $s3_url ]]; then
+      echo "s3_url not provided; EXITING;"
+      exit 1;
+    fi
+
+    s3_user_key=$( kubectl -n s3 get cm s3 -o json | jq -r '.data."s3-user-key"' )
+
     echo Onboarding default partners
     helm -n $NS install esignet-demo-oidc-partner-onboarder mosip/partner-onboarder \
-    --set onboarding.configmaps.s3.s3-host='http://minio.minio:9000' \
-    --set onboarding.configmaps.s3.s3-user-key='admin' \
-    --set onboarding.configmaps.s3.s3-region='' \
+    --set onboarding.configmaps.s3.s3-host="$s3_url" \
+    --set onboarding.configmaps.s3.s3-user-key="$s3_user_key" \
+    --set onboarding.configmaps.s3.s3-region="$s3_region" \
+    --set onboarding.configmaps.s3.s3-bucket-name="$s3_bucket" \
     $ENABLE_INSECURE \
     -f values.yaml \
-    --version $CHART_VERSION
+    --version $CHART_VERSION \
+    --wait --wait-for-jobs
 
-    echo "Reports are moved to S3 under onboarder bucket"
-    echo "Please follow the steps as mentioned in the document link below to configure mock-relying-party-partner:"
-    BRANCH_NAME=$(git symbolic-ref --short HEAD)
-    GITHUB_URL="https://github.com/mosip/esignet-mock-services/blob"
-    FILE_PATH="/README.md"
-    FULL_URL="$GITHUB_URL/$BRANCH_NAME$FILE_PATH#configuration"
+    private_public_key_pair=$(kubectl logs -n $NS job/esignet-demo-oidc-partner-onboarder-demo-oidc | grep -Pzo "(?s)Private and Public KeyPair:\s*\K.*?(?=\s*mpartner default demo OIDC clientId:)" | tr -d '\0' | tr -d '\n')
+    echo Encoded Private and Public Key Pair: $private_public_key_pair
+    kubectl patch secret mock-relying-party-service-secrets -n $NS -p '{"data":{"client-private-key":"'$(echo -n "$private_public_key_pair" | base64 | tr -d '\n')'"}}'
+    kubectl rollout restart deployment -n esignet mock-relying-party-service
+    demo_oidc_clientid=$(kubectl logs -n $NS job/esignet-demo-oidc-partner-onboarder-demo-oidc | grep "mpartner default demo OIDC clientId:" | awk '{sub("clientId:", ""); print $5}')
+    echo mpartner default demo OIDC clientId is: $demo_oidc_clientid
+    kubectl -n esignet set env deployment/mock-relying-party-ui CLIENT_ID=$demo_oidc_clientid
 
 
-    echo -e  "\e[1m\e[4m\e[34m\e]8;\a$FULL_URL\e[0m\e[24m\e]8;;\a"
-
-    echo -e "\e[1mHave you completed the changes mentioned in the onboarding document? (y/n)\e[0m"
-    read answer
-
-    if [[ "$answer" =~ [yY](es)* ]]; then
-      echo -e "\e[1m\e[32mPartners onboarded successfully.\e[0m"
-    else
-      echo -e "\e[1m\e[31mPartner onboarding steps are pending. Please complete the configuration steps for onboarding partner.\e[0m"
-    fi
+    echo Reports are moved to S3 under onboarder bucket
     return 0
   fi
 }
